requirements.txt
google-adk==1.19.0
python-dotenv==1.2.1

